к

+++ src/console/migrations/m191002_151217_banner_localization.php
@@ -17,7 +17,7 @@ class m191002_151217_banner_localization extends Migration
         $this->addTable([
             'id' => $this->bigPrimaryKey(),
             'entity_id' => $this->bigInteger()->notNull(),
-            'locale' => $this->integer()->notNull(),
+            'locale' => $this->bigInteger()->notNull(),
             'image' => $this->string(1024),
             'title' => $this->string(1024),
             'url' => $this->string(1024),
